CheckOrderResponse字段类型int64->int

+++ check.go
@@ -38,7 +38,7 @@ type CheckOrderResponse struct {
 	TransactionID string `json:"transaction_id"` // N 微信显示订单号
 	Status        int    `json:"status"`         // Y 0：未支付，1：支付成功
 	Openid        string `json:"openid"`         // N 支付用户的 OPENID(如果支付则有此字段)
-	TotalFee      int64  `json:"total_fee"`      // N 订单金额
+	TotalFee      int    `json:"total_fee"`      // N 订单金额
 	PaidTime      string `json:"paid_time"`      // N 订单支付时间(如果支付则有此字段)
 	Attach        string `json:"attach"`         // N 用户自定义数据
 	Sign          string `json:"sign"`           // Y 数据签名 详见签名算法, 若签名有误, mch.CheckOrder方法会返回错误
